pipeline/requirements.txt
requests
yt_dlp
easyocr
opencv-python
ffmpeg-python
ultralytics 
torch 
torchvision 
scikit-learn 
pyyaml
python-dotenv
pymongo[srv]==3.12
python-dotenv==1.0.0
dnspython==2.4.2